Move markdown adapter into component base
