+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Application/Components/Announcement Detail/View Model/Default/DefaultAnnouncementDetailViewModelFactory.swift
@@ -1,3 +1,4 @@
+import EurofurenceComponentBase
 import EurofurenceModel
 import Foundation.NSAttributedString
