--- Packages/EurofurenceApplication/Sources/EurofurenceApplication/Extensions/NSAttributedString+trim.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension NSAttributedString {
-	public func attributedStringByTrimming(with charSet: CharacterSet) -> NSAttributedString {
-		let mutableString = NSMutableAttributedString(attributedString: self)
-		mutableString.trim(using: charSet)
-		return NSAttributedString(attributedString: mutableString)
-	}
-}
