--- Packages/EurofurenceApplication/Sources/EurofurenceApplication/Extensions/NSMutableAttributedString+trim.swift
@@ -1,20 +1,0 @@
-import Foundation
-
-extension NSMutableAttributedString {
-	public func trim(using charSet: CharacterSet) {
-		var range = (string as NSString).rangeOfCharacter(from: charSet)
-
-		// Trim leading characters from character set.
-		while range.length != 0 && range.location == 0 {
-			replaceCharacters(in: range, with: "")
-			range = (string as NSString).rangeOfCharacter(from: charSet)
-		}
-
-		// Trim trailing characters from character set.
-		range = (string as NSString).rangeOfCharacter(from: charSet, options: .backwards)
-		while range.length != 0 && NSMaxRange(range) == length {
-			replaceCharacters(in: range, with: "")
-			range = (string as NSString).rangeOfCharacter(from: charSet, options: .backwards)
-		}
-	}
-}
