+++ Packages/EurofurenceApplication/Tests/EurofurenceApplicationTests/Application/Components/Announcement Detail/View Model Tests/AnnouncementDetailViewModelFactoryTestBuilder.swift
@@ -2,6 +2,7 @@ import EurofurenceApplication
 import EurofurenceModel
 import EurofurenceModelTestDoubles
 import Foundation
+import XCTEurofurenceComponentBase
 
 class AnnouncementDetailViewModelFactoryTestBuilder {
