+++ Packages/EurofurenceComponents/Tests/EurofurenceComponentBaseTests/Markdown Rendering/DownMarkdownRendererTest.swift
@@ -1,5 +1,4 @@
-import EurofurenceApplication
-import EurofurenceModel
+import EurofurenceComponentBase
 import XCTest
 
 class DownMarkdownRendererTest: XCTestCase {
