+++ Packages/TestUtilities/Sources/TestUtilities/Random Data Generation/Foundation/NSAttributedString+RandomValueProviding.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension NSAttributedString {
+
+    public static var random: NSAttributedString {
+        return NSAttributedString(string: .random)
+    }
+
+}
